*: move "etcdserver/api/v3alarm"
